Move player-queues logic into mixins; route helper services via public accessors
Stateless loading/tracking/feeding logic moves onto QueueLoader/PlaybackTracker/StreamFeeder mixins over a shared _PlayerQueuesBase, so they operate on the controller's own state directly instead of reaching through self.queues._private. The stateful helper services (MediaResolver/Autoplay/SmartShuffle/ManagedPool) stay composition objects and now reach controller state only through public accessors (queue_data/store_sources/get_playlist_tracks/recency_windows). The per-queue config-entry schema moves to config.py with the controller methods as thin delegators. No behaviour change.
