+++ music_assistant/controllers/player_queues/queue_loader.py
@@ -4,11 +4,13 @@ Queue loading for the Player Queues controller.
 Applies the enqueue option (play/replace/next/add) to a batch of resolved items, loads a single
 media item into the queue, resumes from the play-log when the queue is empty, computes the next
 index, and refills the queue (dynamic managed-pool fill and autoplay fill). Owns no per-queue state;
-it reads and mutates the controller's `PlayerQueueData` records via its owning controller.
+it is mixed into the controller and reads/mutates the controller's `PlayerQueueData` records.
 """
+# ruff: noqa: PLR0915
 
 from __future__ import annotations
 
+import random
 from contextlib import suppress
 from typing import TYPE_CHECKING, cast
 
@@ -21,23 +23,38 @@ from music_assistant_models.enums import (
 from music_assistant_models.errors import (
     MediaNotFoundError,
     MusicAssistantError,
+    PlayerUnavailableError,
 )
 from music_assistant_models.media_items import (
     Album,
+    BrowseFolder,
+    ItemMapping,
+    MediaItemType,
+    PlayableMediaItemType,
+    Playlist,
     Track,
     UniqueList,
+    media_from_dict,
 )
 from music_assistant_models.queue_item import QueueItem
 
+from music_assistant.constants import ATTR_ANNOUNCEMENT_IN_PROGRESS
 from music_assistant.controllers.player_queues.autoplay import (
     AutoplayMode,
 )
+from music_assistant.controllers.player_queues.base import _PlayerQueuesBase
 from music_assistant.controllers.player_queues.constants import (
+    CONF_DEFAULT_ENQUEUE_OPTION_LIVE_SOURCES,
     MANAGED_POOL_MAX,
+)
+from music_assistant.controllers.player_queues.helpers import (
+    handle_play_action,
+    has_dynamic_source,
 )
 from music_assistant.controllers.player_queues.managed_pool import gate_tracks
 from music_assistant.controllers.streams.audio_buffer import AudioBuffer
 from music_assistant.controllers.webserver.helpers.auth_middleware import (
+    get_current_user,
     set_current_user,
 )
 from music_assistant.helpers.throttle_retry import BYPASS_THROTTLER
@@ -46,21 +63,11 @@ if TYPE_CHECKING:
     from music_assistant_models.media_items.metadata import MediaItemImage
     from music_assistant_models.player_queue import PlayerQueue
 
-    from music_assistant.controllers.player_queues.controller import PlayerQueuesController
+    from music_assistant.providers.radio_playlist import RadioPlaylistProvider
 
 
-class QueueLoader:
+class QueueLoaderMixin(_PlayerQueuesBase):
     """Load items into a queue: apply the enqueue option, resolve single items, refill the pool."""
-
-    def __init__(self, queues: PlayerQueuesController) -> None:
-        """
-        Initialize the queueloader.
-
-        :param queues: The owning player queues controller.
-        """
-        self.queues = queues
-        self.mass = queues.mass
-        self.logger = queues.logger
 
     async def _enqueue_with_option(
         self,
@@ -70,7 +77,7 @@ class QueueLoader:
         managed_pool: bool,
     ) -> None:
         """Load queue items into the queue according to the given enqueue option."""
-        queue = self.queues._queue_data[queue_id].queue
+        queue = self._queue_data[queue_id].queue
         if queue.state in (PlaybackState.PLAYING, PlaybackState.PAUSED):
             cur_index = (
                 queue.index_in_buffer
@@ -85,18 +92,18 @@ class QueueLoader:
 
         # handle replace: clear all items and replace with the new items
         if option == QueueOption.REPLACE:
-            await self.queues.load(
+            await self.load(
                 queue_id,
                 queue_items=queue_items,
                 keep_remaining=False,
                 keep_played=False,
                 shuffle=shuffle,
             )
-            await self.queues.play_index(queue_id, 0)
+            await self.play_index(queue_id, 0)
             return
         # handle next: add item(s) in the index next to the playing/loaded/buffered index
         if option == QueueOption.NEXT:
-            await self.queues.load(
+            await self.load(
                 queue_id,
                 queue_items=queue_items,
                 insert_at_index=insert_at_index,
@@ -104,7 +111,7 @@ class QueueLoader:
             )
             return
         if option == QueueOption.REPLACE_NEXT:
-            await self.queues.load(
+            await self.load(
                 queue_id,
                 queue_items=queue_items,
                 insert_at_index=insert_at_index,
@@ -114,14 +121,14 @@ class QueueLoader:
             return
         # handle play: replace current loaded/playing index with new item(s)
         if option == QueueOption.PLAY:
-            await self.queues.load(
+            await self.load(
                 queue_id,
                 queue_items=queue_items,
                 insert_at_index=insert_at_index,
                 shuffle=shuffle,
             )
-            next_index = min(insert_at_index, len(self.queues._queue_data[queue_id].items) - 1)
-            await self.queues.play_index(queue_id, next_index)
+            next_index = min(insert_at_index, len(self._queue_data[queue_id].items) - 1)
+            await self.play_index(queue_id, next_index)
             return
         # handle add: add/append item(s) to the remaining queue items
         if option == QueueOption.ADD:
@@ -130,12 +137,12 @@ class QueueLoader:
             # to the player (and prepared for crossfade), so reshuffling it would swap the
             # upcoming track underneath the player and cause an abrupt, non-crossfaded switch.
             if not queue.shuffle_enabled:
-                add_at_index = len(self.queues._queue_data[queue_id].items) + 1
+                add_at_index = len(self._queue_data[queue_id].items) + 1
             elif queue.state in (PlaybackState.PLAYING, PlaybackState.PAUSED):
                 add_at_index = insert_at_index + 1
             else:
                 add_at_index = insert_at_index
-            await self.queues.load(
+            await self.load(
                 queue_id=queue_id,
                 queue_items=queue_items,
                 insert_at_index=add_at_index,
@@ -146,9 +153,9 @@ class QueueLoader:
             # mark first item as new index
             if queue.current_index is None:
                 queue.current_index = 0
-                queue.current_item = self.queues.get_item(queue_id, 0)
+                queue.current_item = self.get_item(queue_id, 0)
                 queue.items = len(queue_items)
-                self.queues.signal_update(queue_id)
+                self.signal_update(queue_id)
 
     async def _load_item(
         self,
@@ -160,7 +167,7 @@ class QueueLoader:
     ) -> None:
         """Try to load the stream details for the given queue item."""
         queue_id = queue_item.queue_id
-        queue = self.queues._queue_data[queue_id].queue
+        queue = self._queue_data[queue_id].queue
 
         # we use a contextvar to bypass the throttler for this asyncio task/context
         # this makes sure that playback has priority over other requests that may be
@@ -179,7 +186,7 @@ class QueueLoader:
         # loudness
         next_track_from_same_album = (
             next_index is not None
-            and (next_item := self.queues.get_item(queue_id, next_index))
+            and (next_item := self.get_item(queue_id, next_index))
             and (
                 queue_item.media_item
                 and hasattr(queue_item.media_item, "album")
@@ -190,14 +197,14 @@ class QueueLoader:
                 and queue_item.media_item.album.item_id == next_item.media_item.album.item_id
             )
         )
-        current_index = self.queues.index_by_id(queue_id, queue_item.queue_item_id)
+        current_index = self.index_by_id(queue_id, queue_item.queue_item_id)
         if current_index is None:
             previous_track_from_same_album = False
         else:
             previous_index = max(current_index - 1, 0)
             previous_track_from_same_album = (
                 previous_index > 0
-                and (previous_item := self.queues.get_item(queue_id, previous_index)) is not None
+                and (previous_item := self.get_item(queue_id, previous_index)) is not None
                 and previous_item.media_item is not None
                 and hasattr(previous_item.media_item, "album")
                 and previous_item.media_item.album is not None
@@ -256,7 +263,7 @@ class QueueLoader:
         # update queue_item.duration from streamdetails if we got a better value
         if queue_item.streamdetails.duration and not queue_item.duration:
             queue_item.duration = queue_item.streamdetails.duration
-            self.queues.signal_update(queue_id, items_changed=True)
+            self.signal_update(queue_id, items_changed=True)
 
         # pre-initialize the AudioBuffer so audio is ready
         # when the player requests it. For the current/first track this ensures
@@ -284,8 +291,8 @@ class QueueLoader:
 
         Will return None if there are no (more) items in the queue.
         """
-        queue = self.queues._queue_data[queue_id].queue
-        queue_items = self.queues._queue_data[queue_id].items
+        queue = self._queue_data[queue_id].queue
+        queue_items = self._queue_data[queue_id].items
         if not queue_items or cur_index is None:
             # queue is empty
             return None
@@ -307,7 +314,7 @@ class QueueLoader:
             "Filling dynamic tracks for queue %s",
             queue_id,
         )
-        queue = self.queues._queue_data[queue_id].queue
+        queue = self._queue_data[queue_id].queue
         # restore the queue owner's user context so provider filters are respected during this
         # background refill (dynamic-playlist generation honours the current user)
         playback_user = (
@@ -318,28 +325,28 @@ class QueueLoader:
         # weighted per source and recency-gated. fill() already sizes the batch to the pool target;
         # the tail cap below is a defensive ceiling so the unplayed tail never grows past
         # MANAGED_POOL_MAX.
-        pool_tracks = await self.queues._managed_pool.fill(queue_id, is_initial=False)
+        pool_tracks = await self._managed_pool.fill(queue_id, is_initial=False)
         # keep the unplayed tail within the bounded pool size (no current_index => nothing played yet)
         played = 0 if queue.current_index is None else queue.current_index + 1
-        unplayed = max(len(self.queues._queue_data[queue_id].items) - played, 0)
+        unplayed = max(len(self._queue_data[queue_id].items) - played, 0)
         headroom = max(MANAGED_POOL_MAX - unplayed, 0)
         queue_items = [
             QueueItem.from_media_item(queue_id, x) for x in pool_tracks[:headroom] if x.available
         ]
         if not queue_items:
             return
-        await self.queues.load(
+        await self.load(
             queue_id,
             queue_items,
-            insert_at_index=len(self.queues._queue_data[queue_id].items) + 1,
+            insert_at_index=len(self._queue_data[queue_id].items) + 1,
         )
 
     async def _fill_autoplay_tracks(self, queue_id: str) -> None:
         """Fill a Queue with additional tracks based on the configured Autoplay mode."""
-        queue = self.queues.get(queue_id)
+        queue = self.get(queue_id)
         if queue is None or not queue.autoplay_enabled:
             return
-        mode = self.queues._autoplay.resolve_mode(queue_id)
+        mode = self._autoplay.resolve_mode(queue_id)
         self.logger.debug(
             "Filling autoplay tracks (mode: %s) for queue %s", mode.value, queue.display_name
         )
@@ -351,16 +358,16 @@ class QueueLoader:
         set_current_user(playback_user)
         existing_tracks = {
             item.media_item
-            for item in self.queues._queue_data[queue_id].items
+            for item in self._queue_data[queue_id].items
             if isinstance(item.media_item, Track)
         }
         try:
             if mode == AutoplayMode.PLAYLIST:
-                tracks = await self.queues._autoplay.get_playlist_tracks(queue, existing_tracks)
+                tracks = await self._autoplay.get_playlist_tracks(queue, existing_tracks)
             elif mode == AutoplayMode.LIBRARY:
-                tracks = await self.queues._autoplay.get_library_tracks(queue, existing_tracks)
+                tracks = await self._autoplay.get_library_tracks(queue, existing_tracks)
             elif mode == AutoplayMode.SIMILAR:
-                tracks = await self.queues._get_similar_tracks(
+                tracks = await self._get_similar_tracks(
                     queue_id, seed_items=queue.enqueued_media_items
                 )
             else:
@@ -369,11 +376,11 @@ class QueueLoader:
                 # that here to make sure the library fallback still runs.
                 tracks = []
                 with suppress(MusicAssistantError):
-                    tracks = await self.queues._get_similar_tracks(
+                    tracks = await self._get_similar_tracks(
                         queue_id, seed_items=queue.enqueued_media_items
                     )
                 if not tracks:
-                    tracks = await self.queues._autoplay.get_library_tracks(queue, existing_tracks)
+                    tracks = await self._autoplay.get_library_tracks(queue, existing_tracks)
         except MusicAssistantError as err:
             self.logger.warning(
                 "Autoplay failed to fetch tracks for queue %s: %s", queue.display_name, err
@@ -381,7 +388,7 @@ class QueueLoader:
             return
         # route the autoplay batch through the recency engine so a recently-heard track isn't
         # immediately re-added (ungated fallback keeps autoplay going if everything is recent)
-        windows = self.queues._smart_shuffle._windows(queue_id)
+        windows = self._smart_shuffle.windows(queue_id)
         snapshot = await self.mass.music.recency.snapshot(windows, userid=queue.userid)
         tracks = gate_tracks(
             [track for track in tracks if isinstance(track, Track)], snapshot, windows
@@ -390,10 +397,10 @@ class QueueLoader:
         if not queue_items:
             self.logger.info("Autoplay found no new tracks to add for queue %s", queue.display_name)
             return
-        await self.queues.load(
+        await self.load(
             queue_id,
             queue_items,
-            insert_at_index=len(self.queues._queue_data[queue_id].items) + 1,
+            insert_at_index=len(self._queue_data[queue_id].items) + 1,
         )
 
     async def _try_resume_from_playlog(self, queue: PlayerQueue) -> bool:
@@ -430,7 +437,7 @@ class QueueLoader:
                 if not item.uri:
                     continue
                 try:
-                    await self.queues._handle_play_media(queue.queue_id, item)
+                    await self._handle_play_media(queue.queue_id, item)
                     self.logger.info(
                         "Resumed queue %s from playlog (%s)", queue.display_name, match_type
                     )
@@ -440,3 +447,278 @@ class QueueLoader:
                     continue
 
         return False
+
+    @handle_play_action
+    async def _handle_play_media(
+        self,
+        queue_id: str,
+        media: MediaItemType | ItemMapping | str | list[MediaItemType | ItemMapping | str],
+        option: QueueOption | None = None,
+        radio_mode: bool = False,
+        start_item: PlayableMediaItemType | str | None = None,
+        sort_by: str | None = None,
+    ) -> None:
+        """Handle play media without acquiring the queue lock."""
+        # cancel any pending play_index calls for this queue to prevent conflicts
+        self.mass.cancel_timer(f"queue_play_index_{queue_id}")
+        self._set_transitioning(queue_id, False)
+        # we use a contextvar to bypass the throttler for this asyncio task/context
+        # this makes sure that playback has priority over other requests that may be
+        # happening in the background
+        BYPASS_THROTTLER.set(True)
+        if not (queue := self.get(queue_id)):
+            raise PlayerUnavailableError(f"Queue {queue_id} is not available")
+        # always fetch the underlying player so we can raise early if its not available
+        queue_player = self.mass.players.get_player(queue_id, True)
+        assert queue_player is not None  # for type checking
+        if queue_player.extra_data.get(ATTR_ANNOUNCEMENT_IN_PROGRESS):
+            self.logger.warning("Ignore queue command: An announcement is in progress")
+            return
+
+        # save the user requesting the playback (clear it for anonymous playback)
+        playback_user = get_current_user()
+        queue.userid = playback_user.user_id if playback_user else None
+        if playback_user:
+            self.logger.debug(
+                "User %s requested playback.", playback_user.display_name or playback_user.username
+            )
+
+        # a single item or list of items may be provided
+        media_list = media if isinstance(media, list) else [media]
+
+        if radio_mode:
+            # radio_mode is deprecated: a "radio" is now a dynamic radio playlist. Translate each
+            # seed into the radio_playlist provider's URI and enqueue those (resolved to dynamic
+            # playlists that self-manage their refills).
+            self.logger.warning(
+                "radio_mode is deprecated; enqueue a radio_playlist:// dynamic playlist instead"
+            )
+            media_list = [
+                seed_uri
+                if (seed_uri := item if isinstance(item, str) else str(item.uri)).startswith(
+                    "radio_playlist://"
+                )
+                else f"radio_playlist://playlist/{seed_uri}"
+                for item in media_list
+            ]
+            radio_mode = False
+
+        # clear queue if needed
+        if option == QueueOption.REPLACE:
+            self.clear(queue_id, skip_stop=True)
+        # Clear the 'enqueued media item' list when a new queue is requested
+        if option not in (QueueOption.ADD, QueueOption.NEXT):
+            queue.enqueued_media_items.clear()
+
+        # An ADD/NEXT onto a queue that already has dynamic sources keeps feeding its bounded pool;
+        # any other enqueue is not managed here (a REPLACE clears the sources above, and a dynamic
+        # playlist self-manages its own refills).
+        already_managed = bool(queue.sources) and option in (QueueOption.ADD, QueueOption.NEXT)
+        managed_pool = already_managed
+
+        media_items: list[MediaItemType] = []
+        source_items: list[MediaItemType] = []
+        # resolve all media items
+        for item in media_list:
+            try:
+                # parse provided uri into a MA MediaItem or Basic QueueItem from URL
+                media_item: MediaItemType | ItemMapping | BrowseFolder
+                if isinstance(item, str):
+                    media_item = await self.mass.music.get_item_by_uri(item)
+                elif isinstance(item, dict):  # type: ignore[unreachable]
+                    # TODO: Investigate why the API parser sometimes passes raw dicts instead of
+                    # converting them to MediaItem objects. The parse_value function in api.py
+                    # should handle dict-to-object conversion, but dicts are slipping through
+                    # in some cases. This is defensive handling for that parser bug.
+                    media_item = media_from_dict(item)  # type: ignore[unreachable]
+                    self.logger.debug("Converted to: %s", type(media_item))
+                else:
+                    # item is MediaItemType | ItemMapping at this point
+                    media_item = item
+
+                if (
+                    isinstance(media_item, ItemMapping)
+                    and media_item.media_type == MediaType.PLAYLIST
+                ):
+                    # Resolve ItemMapping for a playlist so the full Playlist object
+                    # so we have access to details such as 'is_dynamic'
+                    with suppress(MusicAssistantError):
+                        media_item = await self.mass.music.playlists.get(
+                            media_item.item_id,
+                            media_item.provider,
+                        )
+
+                # Save requested media item to play on the queue so we can use it as a source
+                # for Autoplay. Use FIFO list to keep track of the last 10 played items
+                # Skip ItemMapping and BrowseFolder - only queue full MediaItemType objects
+                if not isinstance(
+                    media_item, (ItemMapping, BrowseFolder)
+                ) and media_item.media_type in (
+                    MediaType.TRACK,
+                    MediaType.ALBUM,
+                    MediaType.PLAYLIST,
+                    MediaType.ARTIST,
+                ):
+                    queue.enqueued_media_items.append(media_item)
+                    if len(queue.enqueued_media_items) > 10:
+                        queue.enqueued_media_items.pop(0)
+                    if isinstance(media_item, Playlist) and media_item.is_dynamic:
+                        # a dynamic playlist/station is always a self-managing dynamic source
+                        source_items.append(media_item)
+
+                # handle default enqueue option if needed
+                if option is None:
+                    # Radio + AudioSource share a single "live_sources" enqueue default —
+                    # both are live infinite streams where REPLACE is almost always the
+                    # right semantic. Other media types use their per-type config key.
+                    if media_item.media_type in (MediaType.RADIO, MediaType.AUDIO_SOURCE):
+                        config_key = CONF_DEFAULT_ENQUEUE_OPTION_LIVE_SOURCES
+                    else:
+                        config_key = f"default_enqueue_option_{media_item.media_type.value}"
+                    config_value = await self.mass.config.get_core_config_value(
+                        self.domain,
+                        config_key,
+                        return_type=str,
+                    )
+                    option = QueueOption(config_value)
+                    if option == QueueOption.REPLACE:
+                        self.clear(queue_id, skip_stop=True)
+
+                # collect media_items to play
+                if isinstance(media_item, Playlist) and media_item.is_dynamic:
+                    # Dynamic playlists/stations supply their own tracks on demand and keep their
+                    # self-managing refill path even inside a managed pool; fetch first batch now.
+                    self.mass.create_task(
+                        self.mass.music.mark_item_played(
+                            media_item, userid=queue.userid, queue_id=queue_id, user_initiated=True
+                        )
+                    )
+                    initial_tracks = await self._media_resolver.get_playlist_tracks(
+                        media_item, start_item=None
+                    )
+                    media_items += initial_tracks
+                elif managed_pool:
+                    # Managed-pool mode: keep the item as a dynamic source; the bounded pool is
+                    # built from these sources below (a finite source rotates its own tracks).
+                    if not isinstance(media_item, (ItemMapping, BrowseFolder)):
+                        source_items.append(media_item)
+                else:
+                    # Convert start_item to string URI if needed
+                    start_item_uri: str | None = None
+                    if isinstance(start_item, str):
+                        start_item_uri = start_item
+                    elif start_item is not None:
+                        start_item_uri = start_item.uri
+                    media_items += await self._media_resolver._resolve_media_items(
+                        media_item,
+                        start_item_uri,
+                        userid=queue.userid,
+                        queue_id=queue_id,
+                        sort_by=sort_by,
+                    )
+
+            except MusicAssistantError as err:
+                # invalid MA uri or item not found error
+                self.logger.warning("Skipping %s: %s", item, str(err))
+
+        # overwrite or append the queue's source items
+        replace_sources = option not in (QueueOption.ADD, QueueOption.NEXT)
+        if replace_sources:
+            self.store_sources(queue, source_items)
+        else:
+            self.store_sources(queue, self._queue_data[queue_id].source_items + source_items)
+        source_items = self._queue_data[queue_id].source_items
+        queue.is_dynamic = has_dynamic_source(source_items)
+
+        if already_managed and not media_items:
+            # new sources were added to an already-active pool: leave the playing pool intact and
+            # let the next top-up incorporate them
+            self.signal_update(queue_id)
+            if queue.current_index is not None and (queue.items - queue.current_index) < 5:
+                task_id = f"fill_dynamic_tracks_{queue_id}"
+                self.mass.call_later(5, self._fill_dynamic_tracks, queue_id, task_id=task_id)
+            return
+
+        # only add valid/available items
+        queue_items: list[QueueItem] = [
+            QueueItem.from_media_item(queue_id, cast("PlayableMediaItemType", x))
+            for x in media_items
+            if x and x.available
+        ]
+
+        if not queue_items:
+            raise MediaNotFoundError("No playable items found")
+
+        # load the items into the queue (managed-pool tracks are already ordered, so don't reshuffle)
+        await self._enqueue_with_option(queue_id, queue_items, option, managed_pool)
+
+    async def _get_similar_tracks(
+        self,
+        queue_id: str,
+        is_initial: bool = False,
+        seed_items: list[MediaItemType] | None = None,
+    ) -> list[Track]:
+        """
+        Fetch tracks similar to the given seeds (autoplay's similar/continuation mode).
+
+        :param queue_id: The queue to fetch tracks for.
+        :param is_initial: True to interleave the base/seed tracks into the result, False to
+            return only similar tracks.
+        :param seed_items: Explicit seed items to base the tracks on. Defaults to the queue's
+            sources; autoplay passes the enqueued media items instead.
+        """
+        queue = self._queue_data[queue_id].queue
+        queue_track_items: list[Track] = [
+            q.media_item
+            for q in self._queue_data[queue_id].items
+            if q.media_item and isinstance(q.media_item, Track)
+        ]
+        source_items = (
+            seed_items if seed_items is not None else self._queue_data[queue_id].source_items
+        )
+        if not source_items:
+            # this may happen during race conditions as this method is called delayed
+            return []
+        self.logger.info(
+            "Fetching similar tracks for queue %s based on: %s",
+            queue.display_name,
+            ", ".join([x.name for x in source_items]),
+        )
+
+        # Get user's preferred provider instances for steering provider selection
+        preferred_provider_instances: list[str] | None = None
+        if (
+            queue.userid
+            and (playback_user := await self.mass.webserver.auth.get_user(queue.userid))
+            and playback_user.provider_filter
+        ):
+            preferred_provider_instances = playback_user.provider_filter
+
+        # Some providers have very deterministic similar-track algorithms for a single track
+        # seed. When continuing from a single track on a refill, seed from the play history
+        # instead so the result keeps varying.
+        if (
+            len(source_items) == 1
+            and source_items[0].media_type == MediaType.TRACK
+            and not is_initial
+            and queue_track_items
+        ):
+            # Helper samples 5 internally; bound the input.
+            seeds: list[MediaItemType] = random.sample(
+                queue_track_items, min(len(queue_track_items), 10)
+            )
+        else:
+            seeds = list(source_items)
+
+        radio_prov = self.mass.get_provider("radio_playlist")
+        if radio_prov is None:
+            return []
+        dynamic_tracks = await cast("RadioPlaylistProvider", radio_prov).get_dynamic_tracks(
+            seeds,
+            include_base_tracks=is_initial,
+            target_size=25,
+            preferred_provider_instances=preferred_provider_instances,
+        )
+        # Drop anything already queued/played
+        queued_set = set(queue_track_items)
+        return [track for track in dynamic_tracks if track not in queued_set]
