+++ tests/controllers/music/test_artist_play_credit.py
@@ -19,7 +19,6 @@ from music_assistant_models.unique_list import UniqueList
 
 from music_assistant.constants import DB_TABLE_ALBUMS, DB_TABLE_ARTISTS, DB_TABLE_PLAYLOG
 from music_assistant.controllers.player_queues import PlayerQueuesController
-from music_assistant.controllers.player_queues.playback_tracker import PlaybackTracker
 from music_assistant.controllers.player_queues.state import PlayerQueueData
 from music_assistant.mass import MusicAssistant
 
@@ -185,11 +184,9 @@ def test_enqueued_album_decision() -> None:
     t3 = Track(item_id="t3", provider="library", name="T3", provider_mappings=set(), album=album_y)
     items = [QueueItem.from_media_item("q1", track) for track in (t1, t2, t3)]
 
-    ctrl = PlayerQueuesController.__new__(PlayerQueuesController)
+    tracker = PlayerQueuesController.__new__(PlayerQueuesController)
     queue = cast("PlayerQueue", Mock(queue_id="q1", enqueued_media_items=[album_x]))
-    ctrl._queue_data = {"q1": PlayerQueueData(queue=queue, items=items)}
-    tracker = PlaybackTracker.__new__(PlaybackTracker)
-    tracker.queues = ctrl
+    tracker._queue_data = {"q1": PlayerQueueData(queue=queue, items=items)}
 
     # first track of the enqueued album -> credit it
     assert tracker._enqueued_album_for_track(queue, items[0], t1) is album_x
